Move GDTCORAssert.h to internal headers
