+++ GoogleDataTransport/GDTCORTests/Unit/Helpers/GDTCORAssertHelper.h
@@ -16,7 +16,7 @@
 
 #import <Foundation/Foundation.h>
 
-#import "GoogleDataTransport/GDTCORLibrary/Public/GoogleDataTransport/GDTCORAssert.h"
+#import "GoogleDataTransport/GDTCORLibrary/Internal/GDTCORAssert.h"
 
 NS_ASSUME_NONNULL_BEGIN
